Added css and file support to webpack, leaflet via bundle
