+++ frontend/package.json
@@ -13,17 +13,26 @@
   "license": "MIT",
   "devDependencies": {
     "@babel/core": "^7.9.0",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/preset-env": "^7.9.5",
     "babel-loader": "^8.1.0",
     "clean-webpack-plugin": "^3.0.0",
+    "css-loader": "^3.5.3",
     "eslint": "^6.8.0",
     "eslint-loader": "^4.0.2",
+    "expose-loader": "^0.7.5",
+    "leaflet": "^1.6.0",
+    "leaflet.featuregroup.subgroup": "^1.0.2",
+    "leaflet.markercluster": "^1.4.1",
+    "style-loader": "^1.2.0",
     "terser-webpack-plugin": "^2.3.5",
+    "url-loader": "^4.1.0",
     "webpack": "^4.42.1",
     "webpack-bundle-tracker": "^0.4.3",
     "webpack-cli": "^3.3.11"
   },
   "dependencies": {
+    "core-js": "^3.6.5",
     "lodash": "^4.17.15"
   }
 }
