+++ frontend/src/map.js
@@ -1,0 +1,2 @@
+import { MapViewPage } from './map/map.mjs'
+window.MapViewPage = MapViewPage
